% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_theta.R
\name{estimation_theta}
\alias{estimation_theta}
\title{Structural estimation for life-cycle model over microsimulated data}
\usage{
estimation_theta(
  theta_0,
  model_function = {     function(theta) theta },
  approach = c("two_step", "one_step"),
  ...
)
}
\arguments{
\item{theta_0}{Initial values for \eqn{\theta} parameter. This can be a
named vector}

\item{model_function}{Function to transform \code{theta} into sample conditions.
See examples.}

\item{approach}{Estimation approach. Either \emph{one_step} or \emph{two_step} (default)}

\item{...}{Additional arguments that should be used by
\code{model_function}}

\item{optim_args}{Arguments that should be used to control \link[stats]{optim}
routine}
}
\description{
User front-end function for indirect inference on microsimulated data.
Estimation is performed using simulated method of moments and
is a particular case of extremum estimator. Theoretical
framework can be found on
\insertCite{gourieroux1993simulation;textual}{mindist}.

Weight matrix \eqn{W} is either assumed to be identity
(\code{approach} = \emph{"one_stage"}) or
estimated with the optimal weight matrix as
proposed by \insertCite{hansen1982large;textual}{mindist}.
In the former case, known as feasible GMM,
(\code{approach} = \emph{"two_stage"})
}
\examples{
\dontrun{

n <- 10000L
ncol <- 3
x <- replicate(ncol, rnorm(n))
y <- 2*x[,2] + runif(n)

df <- data.frame(y = y, x1 = x[,1], x2 = x[,2],
                 x3 = x[,3])

moment_function <- function(theta, ...){
  return(
    data.table::data.table(
      'epsilon' = as.numeric(df$x1*(df$y - cbind(1L, df$x1) \%*\% theta))
    )
  )
}
objective_function <- function(theta, weights = 1L, return_moment = FALSE, ...){
  if (return_moment) return(moment_function(theta, ...))
  if (weights == 1L) return(
  t(moment_function(theta, ...)$epsilon) \%*\%
     moment_function(theta, ...)$epsilon
  )
  return(
    t(moment_function(theta, ...)$epsilon) \%*\%
      weights \%*\%
      moment_function(theta, ...)$epsilon
    )
}

msm1 <- estimation_theta(theta_0 = c(0, 0),
                         model_function = objective_function,
                         approach = "one_step")

}


}
\references{
\insertAllCited{}
}
