% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_jacobian.R
\name{extract_moment}
\alias{extract_moment}
\title{Internal function to extract moments}
\usage{
extract_moment(theta, model_function, ...)
}
\arguments{
\item{theta}{\eqn{\theta} parameter}

\item{model_function}{Function that should be used
to transform \eqn{\theta} parameter into
moment conditions}

\item{...}{Additional arguments for loss function}
}
\value{
Mismatch between theoretical and sample moments
}
\description{
Internal function to extract moments
}
