% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindist.R
\docType{package}
\name{mindist}
\alias{mindist}
\title{mindist: A package for minimum distance estimation}
\description{
This package proposes a series of functions to
perform classical minimum distance and GMM
estimation
}
