% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_weight_matrix.R
\name{optimal_weight_matrix}
\alias{optimal_weight_matrix}
\title{Efficient weight matrix in GMM framework}
\usage{
optimal_weight_matrix(epsilon)
}
\arguments{
\item{epsilon}{Vector of difference between sample and simulated moments}
}
\description{
Implementation of the GMM optimal matrix \insertCite{hansen1982large}{mindist}
in a simulated method of moments framework
}
