% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sargan_test.R
\name{sargan_test}
\alias{sargan_test}
\title{Overidentification test for method of simulated moments}
\usage{
sargan_test(theta_hat, S, model_function = {
    function(theta) theta
}, ...)
}
\arguments{
\item{theta_hat}{Estimated theta parameter}

\item{S}{Estimated covariance matrix}

\item{model_function}{\eqn{g(\theta)} transformation
defining moment conditions}

\item{...}{Additional arguments that should be passed
to \code{model_function}}
}
\value{
List with following elements
}
\description{
Overidentification test for method of simulated moments
}
