% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_function.R
\name{loss_function}
\alias{loss_function}
\title{Compute loss function in a flexible way}
\usage{
loss_function(
  theta,
  prediction_function = {     function(theta) theta },
  weights = 1L,
  verbose = FALSE,
  return_moment = FALSE,
  moments_weights = NULL,
  ...
)
}
\arguments{
\item{theta}{Vector of structural parameters. Assuming a named vector.}

\item{prediction_function}{Function that transforms \eqn{\theta}
into vector of moments. In Newer and MacFadden, denoted \eqn{g(\theta)}}

\item{weights}{Weight matrix \eqn{W} that should be used}

\item{verbose}{Logical value indicating whether we want to print informative
messages to monitor progress}

\item{return_moment}{Logical value indicating whether we want to
return moment matrix (observed, simulation, difference)}

\item{moments_weights}{User-defined weights that should be applied
to reweight moments importance. This is a user choice
independent of using an optimal weight matrix}

\item{...}{Additional arguments that should be used to control
\code{prediction_function} behavior. This function should return a
\code{data.table} object with a variable denoted \code{epsilon} giving the
distance that should be minimized}
}
\value{
A distance between empirical and simulated moments

\link{estimation_theta} for GMM estimation
}
\description{
Compute loss associated with a vector of structural parameters
between some data and a transformation of structural parameters
}
