% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_weight_matrix.R
\name{optimal_weight_matrix}
\alias{optimal_weight_matrix}
\title{Efficient weight matrix in GMM framework}
\usage{
optimal_weight_matrix(epsilon)
}
\arguments{
\item{epsilon}{Vector of difference between sample and simulated moments}
}
\description{
Implementation of the GMM optimal matrix \insertCite{hansen1982large}{wealthyR}
in a simulated method of moments framework
}
\details{
Denoting \eqn{\varepsilon} the difference between sample
and simulated moments, GMM objective function is
known to be
\deqn{\theta = \arg \min_{\theta} \bigg\{ \mathcal{l}(\theta) = \epsilon(\theta)^\intercal W(\theta) \epsilon(\theta) \bigg\}}
The optimal weight matrix is unknown, an estimator is
\deqn{\widehat{W} \big(\widehat{\theta_{(1)}}\big) = \mathbb{E}\bigg(\varepsilon \left(\widehat{\theta_{(1)}} \right) \varepsilon \left( \widehat{\theta_{(1)}}\right)^\intercal \bigg)}
where \eqn{\theta_{(1)}} is a first step GMM estimator
}
